src-tauri/pyproject.toml
[project]
name = "sushi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "you" }
]
requires-python = ">=3.9"
dependencies = [
    "pytauri == 0.8.*",
    "pydantic == 2.*",
    "anyio == 4.*",
    "ijson >= 3.0",
    "watchdog >= 3.0",
    # RAG Pipeline dependencies
    "google-genai >= 1.0.0",
    "faiss-cpu >= 1.7.4",
    "networkx >= 3.0",
    "httpx >= 0.25.0",
    "numpy >= 1.24.0",
]

[project.entry-points.pytauri]
ext_mod = "sushi.ext_mod"

[build-system]
requires = ["setuptools >= 80", "setuptools-scm >= 8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages]
find = { where = ["src-python"] }
